dash_api: Df manipulation factorization, retirements API initialization
