+++ src/apps/services/klima_retirements.py
@@ -1,14 +1,14 @@
-from . import S3, KeyCacheable, final_cached_command, chained_cached_command, single_cached_command
+from . import S3, DfCacheable, final_cached_command, chained_cached_command, single_cached_command
 
 
-class KlimaRetirements(KeyCacheable):
+class KlimaRetirements(DfCacheable):
     """Service for carbon metrics"""
     def __init__(self, commands=[]):
         super(KlimaRetirements, self).__init__(commands)
-
+    
     def getDf(self):
         """Get klima retirements"""
-        return S3().load("raw_polygon_klima_retirements")
+        return S3().load("polygon_klima_retirements")
 
     @single_cached_command()
     def raw(self):
@@ -19,11 +19,6 @@ class KlimaRetirements(KeyCacheable):
     def get(self, _df):
         """Get klima retirements"""
         return self.getDf()
-
-    @chained_cached_command()
-    def daily_agg(self, _df):
-        """Get daily klima retirements"""
-        return S3().load("raw_polygon_klima_retirements_daily")
 
     @final_cached_command()
     def filter_tokens(self, df, tokens):
